# api/openapi-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service
  contact:
    email: maxdanielton@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080
tags:
  - name: payment
    description: Everything about your Payments
paths:
  /deposit:
    post:
      tags:
        - payment
      summary: Add a new deposit
      description: Add a new deposit
      operationId: addDeposit
      requestBody:
        description: Create a new deposit
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/DepositRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/DepositResponse'
        '400':
          description: Invalid input
        '500':
          description: Internal Error
  /withdrawal:
    post:
      tags:
        - payment
      summary: Add a new withdrawal
      description: Add a new deposit
      operationId: addWithdrawal
      requestBody:
        description: Create a new withdrawal
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawalRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/WithdrawalRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
        '400':
          description: Invalid input
        '500':
          description: Internal Error
  /transactions/{id}:
    get:
      tags:
        - payment
      summary: Find transaction by ID
      description: Returns a single transaction
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          description: ID of transaction to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
        '500':
          description: Internal Error
components:
  schemas:
    GatewayDetails:
      type: object
      properties:
        id:
          type: string
          example: gatewayA
        name:
          type: string
          example: fehguy
        callbackUrl:
          type: string
          example: http://localhost:8080/callback
      xml:
        name: GatewayDetails
    Money:
      type: object
      properties:
        amount:
          type: integer
          format: int64
          example: 100
        currency:
          type: string
          example: USD
      xml:
        name: Money
    CardDetails:
      type: object
      properties:
        name:
          type: string
          example: John
        number:
          type: string
          example: 4111111111111111
        type:
          type: string
          example: James
        expiryMonth:
          type: integer
          description: Expiry Month
          format: int64
          example: 12
        expiryYear:
          type: integer
          description: Expiry Year
          format: int64
          example: 2024
        cvv:
          type: string
          example: "123"
      xml:
        name: CardDetails
    Transaction:
      type: object
      properties:
        id:
          type: string
          example: 70cadc76-1eac-4bcd-93dc-8fec928d48d0
        amount:
          $ref: '#/components/schemas/Money'
        cardDetails:
          $ref: '#/components/schemas/CardDetails'
        gatewayDetails:
          $ref: '#/components/schemas/GatewayDetails'
        type:
          type: string
          example: deposit
        status:
          type: string
          example: succeeded
        externalId:
          type: string
          example: 70cadc76-1eac-4bcd-93dc-8fec928d48d0
        createdAt:
          type: string
          example: 2024-09-29 14:36:03.119077077 +0000 UTC
        updatedAt:
          type: string
          example: 2024-09-29 14:36:03.119077077 +0000 UTC
      xml:
        name: Transaction
    DepositRequest:
      required:
        - amount
        - cardDetails
        - gatewayDetails
      type: object
      properties:
        amount:
          $ref: '#/components/schemas/Money'
        cardDetails:
          $ref: '#/components/schemas/CardDetails'
        gatewayDetails:
          $ref: '#/components/schemas/GatewayDetails'
      xml:
        name: DepositRequest
    WithdrawalRequest:
      required:
        - amount
        - cardDetails
        - gatewayDetails
      type: object
      properties:
        amount:
          $ref: '#/components/schemas/Money'
        cardDetails:
          $ref: '#/components/schemas/CardDetails'
        gatewayDetails:
          $ref: '#/components/schemas/GatewayDetails'
      xml:
        name: WithdrawalRequest
    DepositResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: 70cadc76-1eac-4bcd-93dc-8fec928d48d0
        data:
          $ref: '#/components/schemas/DepositRequest'
        message:
          type: string
          example: deposited
        status:
          type: string
          example: succeeded
        processedAt:
          type: string
          example: 2024-09-29 14:36:03.119077077 +0000 UTC
      xml:
        name: DepositResponse
    WithdrawalResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: 70cadc76-1eac-4bcd-93dc-8fec928d48d0
        data:
          $ref: '#/components/schemas/DepositRequest'
        message:
          type: string
          example: deposited
        status:
          type: string
          example: succeeded
        processedAt:
          type: string
          example: 2024-09-29 14:36:03.119077077 +0000 UTC
      xml:
        name: WithdrawalResponse
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: http status code
          format: int64
          example: 500
        message:
          type: string
          example: error
        details:
          type: string
      xml:
        name: WithdrawalResponse
  requestBodies:
    DepositRequest:
      description: Deposit object that needs to be added
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DepositRequest'
        application/xml:
          schema:
            $ref: '#/components/schemas/DepositRequest'
    WithdrawalRequest:
      description: Withdrawal object that needs to be added
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WithdrawalRequest'
        application/xml:
          schema:
            $ref: '#/components/schemas/WithdrawalRequest'